requirements.txt
#Langchain Core
langchain
langchain-Core

#Langchain OpenAI integration
langchain-openai
openai

#Anthropic integration
langchain-anthropic

#Google Gemini (PaLM) integration
langchain-google-genai
google-generativeai

#Hugging Face integration
langchain-huggingface
transformers
huggingface-hub

#Environment variables
python-dotenv

#Machine Learning
scikit-learn
numpy



